Tutorial complete
This is the react front-end UI component. It contains the HTML, CSS and Typescript code from following along the tutorial.

There is a server side project that exposes the REST API endpoints accessed from this project. This project has hard-coded the expectation that the server is running on localhost port 5000.